==> Code/1_scrapGM.ipynb <==
import time
import re
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.common.by import By
import math
import pandas as pd 


def haversine(coord1, coord2):
    lat1, lon1 = [float(i) for i in coord1]
    lat2, lon2 = [float(i) for i in coord2]
    rad=math.pi/180
    dlat= lat2-lat1
    dlon=lon2-lon1
    R=6372.795477598
    a=(math.sin(rad*dlat/2))**2 + math.cos(rad*lat1)*math.cos(rad*lat2)*(math.sin(rad*dlon/2))**2
    distancia=2*R*math.asin(math.sqrt(a))
    return distancia

def get_geocoder(url_location): # gets geographical lat/long coordinates
        try:
            #print(url_location)
            coords = re.search(r"!3d-?\d\d?\.\d{4,8}!4d-?\d\d?\.\d{4,8}",
                            url_location).group()
            coord = coords.split('!3d')[1]
            return tuple(coord.split('!4d'))
        except (TypeError, AttributeError):
            return ("", "")

def scroll_the_page(i):
    try:
        #section_loading = self.driver.find_element_by_class_name("section-loading")
        section_loading = driver.find_element(By.CLASS_NAME, "section-loading")
        while True:
            # if i >= len(self.driver.find_elements(By.CLASS_NAME, "place-result-container-place-link")):
            if i >= len(driver.find_elements_by_xpath('//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[3]/div/a')):
                actions = ActionChains(driver)
                actions.move_to_element(section_loading).perform()
                time.sleep(2)
            else:
                break
    except:
        pass


def Busqueda_20(string_busqueda):
    PATH = '/home/pomilusky/Escriptori/Ironhack/chromedriver'
    query = string_busqueda
    a = dict({})
    options = webdriver.ChromeOptions()
    # options.add_argument("headless")
    options.add_argument("no_sandbox")
    driver = webdriver.Chrome(PATH, options=options)
    driver.get("https://www.google.es/maps/search/"+query.replace(" ", "+")+"/")
    driver.implicitly_wait(5)
    driver.find_element(By.XPATH, '//*[@id="yDmH0d"]/c-wiz/div/div/div/div[2]/div[1]/div[4]/form/div/div/button/span').click()
    n = int(driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[2]/div/div[1]/span/span[2]').text)
    driver.implicitly_wait(5)
    driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[3]/div/a').click()
    name= driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[1]/div[1]/h1/span[1]').text
    time.sleep(5)
    a[name+'_0']= get_geocoder(driver.current_url)
    # print(driver.current_url)
    for i in range(2,n+1):
        try:
            # driver.find_element_by_xpath('//*[@id="omnibox-singlebox"]/div[1]/div[1]/button').click()
            driver.implicitly_wait(5)
            # scroll_the_page(2)
            name= driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[1]/div[1]/h1/span[1]').text
            driver.implicitly_wait(5)
            driver.find_element(By.XPATH,f'//*[@id="sGi9mc-m5SR9c-bottom-pane"]/div/div[1]/div/div/div/div[2]/div[2]/div[{i}]/div[3]/div[1]/div[1]/span[1]').click()
            time.sleep(3)
            a[name+f'_{i}']= get_geocoder(driver.current_url)
            try:
                if i%3 == 0:
                    driver.find_element(By.XPATH,'//*[@id="sGi9mc-m5SR9c-bottom-pane"]/div/div[1]/div/div/div/button[2]/div').click()
                    time.sleep(2)
            except:
                pass

        except:
            print(i)
    driver.quit()
    # print(a)
    return a
    # print
    
    


# Ara em faré una funció que em vagi a una localització i em trobi l'element més proper del tipus que li demani.
def NearSearch(element,coords, n=5):
    if n>20: n = 20
    PATH = '/home/pomilusky/Escriptori/Ironhack/chromedriver'
    query = element.replace(" ", "+")
    a = dict({})
    lat, lon = coords
    options = webdriver.ChromeOptions()
    # options.add_argument("headless")
    options.add_argument("no_sandbox")
    driver = webdriver.Chrome(PATH, options=options)
    driver.get(f"https://www.google.com/maps/search/{query}/@{lat},{lon},12z")
    driver.implicitly_wait(5)
    driver.find_element(By.XPATH, '//*[@id="yDmH0d"]/c-wiz/div/div/div/div[2]/div[1]/div[4]/form/div/div/button/span').click()
    driver.implicitly_wait(5)
    driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[3]/div/a').click()
    name= driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[1]/div[1]/h1/span[1]').text
    time.sleep(5)
    crds = get_geocoder(driver.current_url)
    dist = haversine(coords, crds)
    
    a[1]= {'nombre':name,'location':{'type': 'Point', 'coordinates':list(crds)}, 'distance': dist, 'Entidad': element}

    # print(driver.current_url)
    for i in range(2,n+1):
        try:
            # driver.find_element_by_xpath('//*[@id="omnibox-singlebox"]/div[1]/div[1]/button').click()
            driver.implicitly_wait(5)
            # scroll_the_page(2)
            name= driver.find_element(By.XPATH,'//*[@id="pane"]/div/div[1]/div/div/div[2]/div[1]/div[1]/div[1]/h1/span[1]').text
            driver.implicitly_wait(5)
            driver.find_element(By.XPATH,f'//*[@id="sGi9mc-m5SR9c-bottom-pane"]/div/div[1]/div/div/div/div[2]/div[2]/div[{i}]/div[3]/div[1]/div[1]/span[1]').click()
            time.sleep(3)
            crds = get_geocoder(driver.current_url)
            dist = haversine(coords, crds)
            a[i]= {'nombre':name,'location':{'type': 'Point', 'coordinates':list(crds)}, 'distance': dist, 'Entidad': element}

            try:
                if i%3 == 0:
                    driver.find_element(By.XPATH,'//*[@id="sGi9mc-m5SR9c-bottom-pane"]/div/div[1]/div/div/div/button[2]/div').click()
                    time.sleep(2)
            except:
                pass

        except:
            print(i)
    driver.quit()
    # print(a)
    return pd.DataFrame(dict(sorted(a.items() , key = lambda x: x[1]['distance']))).T

df = pd.read_csv('../Data/castilla_leon/toMongo.csv', index_col='Unnamed: 0')

df[['Entidad Singular de Población','location']].head()

i = 5
list(df[df['Entidad Singular de Población'] == 'NAVALUENGA'].location)

to_scrap = dict({'NAVALUENGA':(40.411111111111,-4.7077777777778)})
to_scrap

for k,v in to_scrap.items():
    try:
        print(k, v)
        things_2_look_4 = ['supermercado','escuela','hospital','centro de salud', 'gasolinera']
        lista_df = []
        for i in things_2_look_4:
            df = NearSearch(i,v)
            lista_df.append(df)
        new_df = pd.concat(lista_df, axis=0, join="outer", ignore_index=True, keys=None, levels=None, names=None, verify_integrity=False, copy=False)
        a = k.replace(' ','_')
        new_df.to_csv(f'../Data/castilla_leon/{a}.csv')
    except: pass

df


pd.DataFrame(a)

mec = Busqueda_20('Hospitales Zamora')
mec

driver = webdriver.Chrome(executable_path='/home/pomilusky/Escriptori/Ironhack/chromedriver')
