==> Code/1_Weather_scrap.ipynb <==
import pandas as pd
import requests
import re
from bs4 import BeautifulSoup
#import selenium as sel
import numpy as np
from selenium import webdriver
from selenium.webdriver.common.keys import Keys


df = pd.read_csv('../Data/castilla_leon/main.csv', index_col='Unnamed: 0')

def get_weather(pob):
    try:
        url = 'https://es.weatherspark.com/'
        options = webdriver.ChromeOptions()
        options.add_argument("headless")
        options.add_argument("no_sandbox")
        driver = webdriver.Chrome(executable_path='/home/pomilusky/Escriptori/Ironhack/chromedriver',chrome_options=options)
        driver.get(url)
        driver.implicitly_wait(2)
        a = driver.find_element_by_xpath('//*[@id="LiveSearch-FrontPage-InputContainer"]/input[3]')
        a.send_keys(pob)
        driver.find_element_by_xpath('//*[@id="LiveSearch-FrontPage-InputContainer"]/div/a[1]').click()
        driver.implicitly_wait(2)
        driver.find_element_by_xpath('//*[@id="accept-choices"]').click()
        # URL = driver.current_url
        # print(URL)
        texto = driver.find_element_by_xpath('//*[@id="Report-Content"]/div[1]/p[1]').text
        driver.quit()
        return texto
    except:
        return 'Not found'

def cambio_enti(s):
    llista = s.lower().split(' ')
    n_llista = [i.capitalize() if i != 'de' and i != 'del' else i for i in llista ]
    s = ' '.join(n_llista)
    if '-' in s:
        llista = s.lower().split('-')
        n_llista = [i.capitalize() if i != 'de' and i != 'del' else i for i in llista ]
        s = ' '.join(n_llista)
    if '(la)' in s:
        s = 'La '+s.replace('(la)','').strip()
    if '(el)' in s:
        s = 'El '+s.replace('(el)','').strip()
    if '(los)' in s:
        s = 'Los '+s.replace('(los)','').strip()
    if '(las)' in s:
        s = 'Las '+s.replace('(las)','').strip()   

    if 'de La ' in s:
        s = s.replace('de La ', 'de la ')
    if 'de Las ' in s:
        s = s.replace('de Las ', 'de las ')
    if 'de Los ' in s:
        s = s.replace('de Los ', 'de los ')
    if ' Y ' in s:
        s = s.replace(' Y ', ' y ')
    if ' O ' in s:
        s = s.replace(' O ', ' o ')
    return s

df['Entidad Singular de Población'] = df['Entidad Singular de Población'].apply(cambio_enti)

df['Entidad Singular de Población'] = df['Entidad Singular de Población'].apply(get_weather)

df.tail()

Serie =df['Entidad Singular de Población']

Serie

df_1 = pd.read_csv('../Data/castilla_leon/main.csv', index_col='Unnamed: 0')

df_1['Weather_info']=Serie

df_1

df_1.to_csv('../Data/castilla_leon/main.csv')
