==> Code/2_geoqueries.ipynb <==
import pymongo
import pandas as pd
import numpy as np

cliente=pymongo.MongoClient('mongodb://localhost:27017')  # cliente , la conexion por defecto
db=cliente.EmptiedSpain   # llamada a la base de datos


n_pueblo = 'ALBORNOS'
info_pueblo=db.emptiedSpain.find({'Pueblo_desde':n_pueblo})

df = pd.DataFrame(info_pueblo)

def cleanNans(dic):
    a = dic['coordinates']
    if np.nan in a: 
        dic['coordinates'] = [0,0]
    return dic

df.location = df.location.apply(cleanNans)

df.head()

db.emptiedSpain.create_index([('location', '2dsphere')])


def find_near(geopoint, radio=1000):
    return db.emptiedSpain.find({'$and':[{'location':{
        '$near': {'$geometry': geopoint, '$maxDistance':radio}
    }},{'Entidad': 'Hospital'}]})

pueblos=db.emptiedSpain.find({'Entidad': 'Entidad de población'})

df_p = pd.DataFrame(pueblos)

i = 1050
a = df_p.location[i]
df_p.iloc[i]

pd.DataFrame(find_near(a, radio=600000))
