==> Code/2_mergePob&Mun.ipynb <==
import pandas as pd

df_muns = pd.read_csv('../Data/castilla_leon/ParaBD/Total.csv', index_col = ['Unnamed: 0']).reset_index(drop=True)

df_muns.rename(columns={'new_col':'partido_gov'}, inplace=True)

df_muns.drop(columns=['Municipios','Provincia'], inplace=True)
df_muns.head(2) # nos quedamos solo con las columnas con las que haremos append con el otro dataframe

len(df_muns.CP.unique())

df_pob = pd.read_csv('../Data/castilla_leon/main.csv', index_col=['Unnamed: 0']).reset_index(drop=True)

def getCN(C_esp, provincia):
    if provincia in ['Salamanca','Palencia','Soria','León','Segovia','Valladolid','Zamora']:
        return int(str(C_esp)[:5])
    elif provincia in ['Burgos','Ávila']:
        return int(str(C_esp)[:4])

df_pob['CP']  = df_pob.apply(lambda x: getCN(x['Código ESP'], x['Provincia']),axis=1)

df_pob.sample(2)

df_pob.shape

df_tosafe = df_pob.merge(df_muns, on = 'CP')

df_tosafe.rename(columns={'Total':'pob_tot_muni'}, inplace=True)

df_tosafe.to_csv('../Data/castilla_leon/main.csv')
