==> Code/2_mongoDBqueries.ipynb <==
import pandas as pd
import numpy as np
import re

df = pd.read_csv('../Data/castilla_leon/main.csv', index_col='Unnamed: 0').reset_index(drop=True)

df[df['Entidad Singular de Población']=='NAVALUENGA']

df_mec = pd.read_csv('../Data/castilla_leon/mec.csv', index_col='Unnamed: 0').reset_index(drop=True)

df['population'] = df_mec.population

df.tail()

def coord_clean0(s):
    if type(s) != float:
        l = s.lstrip('(').rstrip(')').split(',')
        return tuple(float(i) for i in l)[0]
    else:
        return np.nan
def coord_clean1(s):
    if type(s) != float:
        l = s.lstrip('(').rstrip(')').split(',')
        return tuple(float(i) for i in l)[1]
    else:
        return np.nan

# df_copy.coord.unique()

df['lat'] = df.coord.apply(coord_clean0)
df['lon'] = df.coord.apply(coord_clean1)
df.drop(columns=['coord'], inplace=True)

df.tail(1)

def regex_temp(s):
    l = re.findall('(-\d{1,2}|\d{1,2})',s)
    if l != []:
        return l
    else:
        return[np.nan,np.nan,np.nan,np.nan]
# regex_temp(a)  

df['T_minMedia'] = df.Weather_info.apply(lambda x: regex_temp(x)[0])
df['T_maxMedia'] = df.Weather_info.apply(lambda x: regex_temp(x)[1])
df['T_minExtrema'] = df.Weather_info.apply(lambda x: regex_temp(x)[2])
df['T_maxExtrema'] = df.Weather_info.apply(lambda x: regex_temp(x)[3])

df.head()

df.fillna({'lat':0.0,'lon':0.0}, inplace=True)

def geopoint(lat, lon):
    if lat != np.nan and lon != np.nan:
        return {'type': 'Point', 'coordinates': [lon, lat]}
    else: 
        return {'type': 'Point', 'coordinates': [0.0, 0.0]}

# De cara a poder aplicar geoqueries creamos putos de location que entraran como diccionarios,
df['location'] = df.apply(lambda x: geopoint(x.lat, x.lon), axis=1)
df.drop(columns=['lat','lon'], inplace=True)

df.Weather_info[9]

df['Código ESP']=df['Código ESP'].apply(lambda x: str(x))
df.drop(columns=['CP'], inplace=True)
df.pob_tot_muni = df.pob_tot_muni.apply(lambda x: float(x))

df.to_csv('../Data/castilla_leon/toMongo.csv')

# Entramos tambien datos de hospitales

df_h = pd.read_csv('../Data/castilla_leon/salud/hospitals.csv', index_col=['OBJECTID'])
df_h = df_h.reset_index(drop=True).drop(columns=[ 'TELEFONO2', 'TELEFAX','CODFU', 'DEPENDENCIA_FUNCIONAL',
       'ACREDOCENT', 'ESCOMPLE', 'FORCOMPLE', 'CODIDCOM', 'ALTA', 'CERRADO',
       'CAPITAL', 'CIERREFECH', 'CONCIERTO', 'EMAIL', 'TAC', 'RM', 'GAM',
       'HEM', 'ASD', 'LIT', 'BCO', 'ALI', 'SPECT', 'PET', 'MAMOS', 'DO',
       'DIAL','CalidadGeocodificacion','CODPAT','CODFI','CODAUTO'])
df_h.rename(columns={'X': 'lon','Y':'lat'}, inplace=True)
df_h = df_h[df_h.COMUNIDADES=='CASTILLA Y LEÓN']

df_h.head(3)

df_h['location'] = df_h.apply(lambda x: geopoint(x.lat, x.lon), axis=1)
df_h.drop(columns=['lat','lon'], inplace=True)

# df_h.info()
list_int64 = ['CODCNH','CODMU','CODPROV','CODPOSTAL','NCAMAS']
for i in list_int64:
    df_h[i] = df_h[i].apply(lambda x: float(x))

df.location.iloc[0]

# Creacion de la base de datos 

import pymongo

cliente=pymongo.MongoClient()  # llamar a la clase MongoClient

db=cliente.EmptiedSpain           # crear una base de datos

cliente.list_database_names()

# {'type': 'Point', 'coordinates': [-73.987794, 40.740512]}}

colec=db.emptiedSpain   # crea coleccion
db.list_collection_names()

def to_apply(a):
    a['coordinates'][0] = a['coordinates'][0]*-1
    return a
df.location = df.location.apply(to_apply)
df.drop(columns=['pob_tot_muni.1','partido_gov_y'], inplace=True)

# Vamos a crear los diccionarios, i a insertarlos en nuestra base de datos,
colec.drop()
for i in range(df.shape[0]):
    diccio = dict(df.iloc[i])
    # vamos a alterar este diccionario:
    diccio['_id'] = diccio['Código ESP']
    diccio['Entidad'] = 'Entidad de población'
    to_insert = {k:v for k,v in diccio.items() if k not in ['Código ESP']}
    colec.insert_one(to_insert)


lista_municipios = list(df.Municipio.unique())

# Vamos a crear los diccionarios ahora de los hospitales
for i in range(df_h.shape[0]):
    # print(df_h.iloc[i])
    diccio = dict(df_h.iloc[i])
    # vamos a alterar este diccionario:
    diccio['_id'] = diccio['CODCNH']
    diccio['Entidad'] = 'Hospital'
    to_insert = {k:v for k,v in diccio.items() if k not in ['CODCNH']}
    colec.insert_one(to_insert)

def corr_location(s):
    s = s.rstrip('}').lstrip('{')
    l = [float(i) for i in re.findall('-?\d{1,2}\.\d+',s)]
    return {'type':'Point','coordinates': l[::-1]}


pueblos = ['ADANERO','ALBORNOS','BASARDILLA','FORMARIZ','PINILLA_DE_FERMOSELLE','RINCONADA','SAN_BLAS','SAN_CIBRIÁN','TORRALBA_DEL_MORAL','NAVALUENGA']
for pueb in pueblos:
    df_p = pd.read_csv(f'../Data/castilla_leon/{pueb}.csv', index_col=['Unnamed: 0']).reset_index(drop=True)
    df_p.location = df_p.location.apply(corr_location)
    for i in range(df_p.shape[0]):
        diccio = dict(df_p.iloc[i])
        # vamos a alterar este diccionario:
        diccio['Pueblo_desde'] = pueb
        to_insert = diccio
        colec.insert_one(to_insert)
