==> Code/3_folium.ipynb <==
import folium
import pandas as pd

def coords_froms(s):
    list = s.replace('{','').replace('}','').replace('[','').replace(']','').split(':')
    a =[float(i) for i in list[2].replace("'",'').split(',')]
    return a

# guarda el mapa
# mapa.save('../mapa.html')

df_main = pd.read_csv('../Data/castilla_leon/main.csv')

df_ADANERO = pd.read_csv('../Data/castilla_leon/ADANERO.csv', index_col='Unnamed: 0')

df_ADANERO.head(2)

centro = df_main.iloc[0].coord

centro = [float(i.replace('(','').replace(')','')) for i in  centro.split(',')]

centro[1] = -1*centro[1]

mapa=folium.Map(location=centro,
                tiles='openstreetmap', zoom_start=12)

df_ADANERO.shape[0]

df_main.iloc[0]

# Para adanero
mapa=folium.Map(location=[40.945833333333, -4.6066666666667],
                tiles='openstreetmap', zoom_start=12)
informacion = 'info aqui'
folium.CircleMarker([40.945833333333, -4.6066666666667], popup=informacion,tooltip='<strong>Información del pueblo</strong>', radius=10,fill_color='green', color = 'green', 
icon=folium.Icon(color='green',prefix='glyphicon',icon='on')).add_to(mapa)
# marker_cluster = folium.MarkerCluster().add_to(mapa)
for i in range(df_ADANERO.shape[0]):
    point = df_ADANERO.iloc[i]
    st = point['nombre']
    dt = point['distance']
    en = point['Entidad']
    html = f'''<b>{st}</b><br>
            Distancia: {dt:.2f}km<br>
            Entidad: {en}'''

    iframe = folium.IFrame(html,
                        width=100,
                        height=100)

    popup = folium.Popup(iframe,
                        max_width=100)

    folium.Marker(coords_froms(point['location']),popup=popup, tooltip='<strong>Info Entidad</strong>').add_to(mapa)
#     folium.TileLayer('Stamen Terrain').add_to(mapa)
#     folium.TileLayer('Stamen Toner').add_to(mapa)


mapa
